requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-multipart==0.0.6

# Database (SQLite)
sqlalchemy==2.0.25
aiosqlite==0.19.0
alembic==1.13.1

# Document Processing
docling==1.0.0
PyMuPDF==1.23.21
pdfplumber==0.10.4
pytesseract==0.3.10
python-docx==1.1.0
openpyxl==3.1.2
Pillow==10.2.0
pdf2image==1.17.0

# AI/ML Libraries
langchain==0.1.4
langchain-community==0.0.16
langchain-openai==0.0.5
langgraph==0.0.20
sentence-transformers==2.3.1
transformers==4.37.2
torch>=2.2.2
spacy==3.7.2

# Vector Database
chromadb==0.4.22

# OpenAI & Anthropic
openai==1.10.0
anthropic==0.8.1

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
loguru==0.7.2
httpx==0.26.0
aiofiles==23.2.1

# Report Generation
reportlab==4.0.9
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0
kaleido==0.2.1
xlsxwriter==3.1.9
jinja2==3.1.3

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
